Reformating the script to manually align two images in 2D
